backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
email-validator==2.2.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Pydantic for data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Environment and Configuration
python-dotenv==1.0.1

# Development and Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-postgresql==6.1.1

# Code Quality
black==24.8.0
isort==5.13.2
mypy==1.11.2
pre-commit==3.8.0

# Monitoring and Logging
structlog==24.4.0
prometheus-client==0.21.0

# HTTP Client
aiohttp==3.10.8

# Date and Time
python-dateutil==2.9.0.post0
